+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 - **NEW:** Flutterby views now have a `tag` helper method available that can generate HTML tags programatically.
 - **NEW:** Flutterby views now have a `link_to` helper method available that renders link tags. You can use a URL string as the link target, eg. `link_to "Home", "/"`, or any Flutterby node, eg. `link_to "Blog", blog_node`.
+- **NEW:** Flutterby views now have a `debug` helper that will dump its argument's YAML representation into a `<pre>` HTML tag (similar to Rails.)
 
 
 ### 0.3.1 (2017-01-15)